streamlit_requirements.txt
streamlit>=1.31.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=8.0.0 